Promote PanelResizeDirective out of label-view and document its scope
The directive was general-purpose, specced and correct, but lived in
`components/label-view/` with only its two label-view consumers — which is
why every other divider in the app was hand-rolled instead: the shared thing
was invisible.

Move it (with its spec) to `app/directives/` beside `no-focus-steal`, and
document the applicability boundary in its docstring, since "seven duplicate
drags" does not survive contact with the code:

- find-view's two dividers are the directive's exact shape, but deliberately
  do zero per-pixel change detection (they set a CSS custom property and bind
  nothing). A template-bound output *is* a CD notification, so adopting there
  would add a CD pass per mousemove to the heaviest view.
- browse-view's right divider snapshots a DOM-measuring `panelMin()` at
  mousedown; binding it to `[minWidth]` would re-measure every CD cycle.
- browse-bin-popup's metadata divider is an inverted PointerEvent *delta*
  drag, browse-view's row divider maps vertical motion onto a width, and
  browse-minimap's handle is a 2-D corner resize. None are this shape.

Add the `[maxWidth]` input the two browse-view side dividers would need (the
only genuinely clean adoptions), with spec coverage for the cap and for the
inverted-range case, so a future adopter is not blocked on reshaping it.

No behavioural change to the two existing consumers.
